requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyyaml==6.0.1
python-multipart==0.0.6
pydantic[email]

# Database Drivers
psycopg2-binary==2.9.9
pymysql==1.1.0

# Build Dependencies
nuitka
ordered-set==4.1.0

# TUI and Desktop
rich==13.7.0
pywebview==5.0.0


